second to last merge or something

+++ GXPEngine/CoolScaryGame/Utility/ColorSprite.cs
@@ -9,7 +9,7 @@ namespace CoolScaryGame
 {
     public class ColorSprite : Sprite
     {
-        public ColorSprite(int width, int height, uint color) : base("UI/whitePixel.png", true, false)
+        public ColorSprite(float width, float height, uint color) : base("UI/whitePixel.png", true, false)
         {
             this.color = color;
             this.scaleX = width;
